Fix circular dep and remove reference to process.env
